+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Misc:
 - **PHP_CodeSniffer** Provide Sider's recommended ruleset [#2084](https://github.com/sider/runners/pull/2084)
 - Remove *removed* Go tools from `sider.yml` [#2099](https://github.com/sider/runners/pull/2099)
 - Add `Runners::ConfigGenerator` class [#2029](https://github.com/sider/runners/pull/2029)
+- **Metrics FileInfo** Modify target file generation [#2100](https://github.com/sider/runners/pull/2100)
 
 ## 0.43.0